pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpe-automation"
version = "1.0.0"
description = "HPE ProLiant Windows Server ISO Automation Suite"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Platform Engineering Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "jinja2>=3.1.0",
    "colorlog>=6.8.0",
    "paramiko>=3.4.0",
    "psutil>=5.9.0",
    "cryptography>=41.0.0",
    "pywinrm>=0.4.0",
    "xmltodict>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "radon>=6.0.0",
    "bandit>=1.7.0",
    "safety>=3.0.0",
    "mypy>=1.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
build-iso = "automation.cli.build_iso:main"
update-firmware = "automation.cli.update_firmware_drivers:main"
patch-windows = "automation.cli.patch_windows_security:main"
deploy-server = "automation.cli.deploy_to_server:main"
monitor-install = "automation.cli.monitor_install:main"
opsramp = "automation.cli.opsramp_integration:main"
maintenance-mode = "automation.cli.maintenance_mode:main"
generate-uuid = "automation.cli.generate_uuid:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["automation*"]

[tool.ruff]
target-version = "py39"
src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["automation"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
